- 패키지 정리 - 매치히스토리뷰에서 디테일 창 띄우기 기능 추가

+++ src/main/java/net/boardrank/boardgame/ui/currentmatch/MatchDetailModifyDialog.java
@@ -1,4 +1,4 @@
-package net.boardrank.boardgame.ui;
+package net.boardrank.boardgame.ui.currentmatch;
 
 import com.vaadin.flow.component.UI;
 import com.vaadin.flow.component.button.Button;
@@ -9,6 +9,7 @@ import com.vaadin.flow.component.orderedlayout.FlexComponent;
 import com.vaadin.flow.component.orderedlayout.VerticalLayout;
 import lombok.extern.slf4j.Slf4j;
 import net.boardrank.boardgame.domain.GameMatch;
+import net.boardrank.boardgame.domain.GameMatchStatus;
 import net.boardrank.boardgame.service.GameMatchService;
 
 @Slf4j
@@ -27,6 +28,13 @@ public class MatchDetailModifyDialog extends Dialog {
         this.gameMatch = gameMatch;
 
         btn_removeMatch = new Button("Match 삭제하기", event -> {
+            if(gameMatch.getGameMatchStatus().equals(GameMatchStatus.resultAccepted)) {
+                Notification notification = new Notification("해당 Match는 삭제할 수 없습니다.(Match Closed)");
+                notification.setDuration(1500);
+                notification.open();
+                return;
+            }
+
             gameMatchService.removeMatch(gameMatch);
 
             Notification notification = new Notification("Match가 삭제되었습니다.");
